* new style ControlAttributes

+++ Application/src/tracker/tracking/Categorize.cpp
@@ -1973,7 +1973,7 @@ std::shared_ptr<PPFrame> cache_pp_frame(const Frame_t& frame, const std::shared_
             }
 
 #ifndef NDEBUG
-            print("Deleting ",std::distance(start, end)," items from frame cache, which are farther away than ",end != frames_in_cache.end() ? std::get<0>(*end) : -1," from the mean of ",(minimum_range + (maximum_range - minimum_range) / 2.0)," (",_frame_cache.size()," size) ");
+            print("Deleting ",std::distance(start, end)," items from frame cache, which are farther away than ",end != frames_in_cache.end() ? int64_t(std::get<0>(*end)) : -1," from the mean of ",(minimum_range + (maximum_range - minimum_range) / 2.0)," (",_frame_cache.size()," size) ");
 #endif
             _frame_cache = erase_indices(_frame_cache, indices);
             _delete += indices.size();
